Add AR trajectory-replay inference + track data subpackage
Inference (open-loop trajectory replay):
- infer/replay.py: prime with GT history, feed recorded actions, AR-generate
  the rest (latent-space core, CPU-testable)
- data/decode.py: Wan-VAE latent->RGB decoder (inverse of encode.py)
- scripts/replay_ar.py: CLI -> side-by-side GT|PRED videos + MSE/PSNR summary
- tests/test_replay.py: replay plumbing test (dummy backbone, CPU)

Configs:
- configs/training/ar_{wan,cosmos}_droid.py: DROID training configs

.gitignore: anchor /data/, /test/, /tests/ so package code under
openworld/autoregressive/{data,tests}/ is tracked (previously the unanchored
patterns hid the whole data subpackage and the tests).

Also: docs + smoke/validate script updates.

+++ openworld/autoregressive/distill/scheduler.py
@@ -41,15 +41,18 @@ class FlowMatchScheduler:
         return s * sigma / (1 + (s - 1) * sigma)
 
     # -- noising ---------------------------------------------------------
+    # sigma is built in fp32; cast to the operand dtype so a bf16 latent stays
+    # bf16 (otherwise the multiply promotes to fp32 and leaks fp32 into the loss
+    # graph, breaking the bf16 backward).
     def add_noise(self, x0: torch.Tensor, eps: torch.Tensor, sigma: torch.Tensor) -> torch.Tensor:
-        sigma = sigma.view(-1, *([1] * (x0.ndim - 1)))
+        sigma = sigma.view(-1, *([1] * (x0.ndim - 1))).to(x0.dtype)
         return (1 - sigma) * x0 + sigma * eps
 
     def velocity_target(self, x0: torch.Tensor, eps: torch.Tensor) -> torch.Tensor:
         return eps - x0
 
     def x0_from_velocity(self, x_sigma: torch.Tensor, v_pred: torch.Tensor, sigma: torch.Tensor) -> torch.Tensor:
-        sigma = sigma.view(-1, *([1] * (x_sigma.ndim - 1)))
+        sigma = sigma.view(-1, *([1] * (x_sigma.ndim - 1))).to(x_sigma.dtype)
         return x_sigma - sigma * v_pred
 
     def step_sigma(self, i: int, device, batch: int) -> torch.Tensor:
